+++ sdk/src/main/java/ly/count/android/sdk/messaging/CountlyPush.java
@@ -8,7 +8,6 @@ import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -33,6 +32,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -52,12 +52,13 @@ public class CountlyPush {
     public static final String CHANNEL_ID = "ly.count.android.sdk.CountlyPush.CHANNEL_ID";
     public static final String SECURE_NOTIFICATION_BROADCAST = "ly.count.android.sdk.CountlyPush.SECURE_NOTIFICATION_BROADCAST";
     public static final String COUNTLY_BROADCAST_PERMISSION_POSTFIX = ".CountlyPush.BROADCAST_PERMISSION";
+    public static final String ALLOWED_PACKAGE_NAMES = "allowed_package_names";
+    public static final String ALLOWED_CLASS_NAMES = "allowed_class_names";
 
     private static Application.ActivityLifecycleCallbacks callbacks = null;
     private static Activity activity = null;
 
-    private static Countly.CountlyMessagingMode mode = null;
-    private static Countly.CountlyMessagingProvider provider = null;
+    private static CountlyConfigPush countlyConfigPush = null;
 
     static Integer notificationAccentColor = null;
 
@@ -229,10 +230,10 @@ public class CountlyPush {
     public static class ConsentBroadcastReceiver extends BroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent broadcast) {
-            if (mode != null && provider != null && getPushConsent(context)) {
-                String token = getToken(context, provider, Countly.sharedInstance().L);
+            if (countlyConfigPush.mode != null && countlyConfigPush.provider != null && getPushConsent(context)) {
+                String token = getToken(context, countlyConfigPush.provider, Countly.sharedInstance().L);
                 if (token != null && !"".equals(token)) {
-                    onTokenRefresh(token, provider);
+                    onTokenRefresh(token, countlyConfigPush.provider);
                 }
             }
         }
@@ -367,10 +368,7 @@ public class CountlyPush {
             return Boolean.FALSE;
         }
 
-        Intent pushActivityIntent = new Intent(context.getApplicationContext(), CountlyPushActivity.class)
-            .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        pushActivityIntent.setPackage(context.getApplicationContext().getPackageName());
-        pushActivityIntent.putExtra(EXTRA_INTENT, actionIntent(context, notificationIntent, msg, 0));
+        Intent pushActivityIntent = createPushActivityIntent(context, msg, notificationIntent, 0, CountlyPush.countlyConfigPush.allowedIntentClassNames, CountlyPush.countlyConfigPush.allowedIntentPackageNames);
 
         final Notification.Builder builder = (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O ? new Notification.Builder(context.getApplicationContext(), CHANNEL_ID) : new Notification.Builder(context.getApplicationContext()))
             .setAutoCancel(true)
@@ -397,10 +395,7 @@ public class CountlyPush {
         for (int i = 0; i < msg.buttons().size(); i++) {
             Button button = msg.buttons().get(i);
 
-            pushActivityIntent = new Intent(context.getApplicationContext(), CountlyPushActivity.class)
-                .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            pushActivityIntent.setPackage(context.getApplicationContext().getPackageName());
-            pushActivityIntent.putExtra(EXTRA_INTENT, actionIntent(context, notificationIntent, msg, i + 1));
+            pushActivityIntent = createPushActivityIntent(context, msg, notificationIntent, i + 1, CountlyPush.countlyConfigPush.allowedIntentClassNames, CountlyPush.countlyConfigPush.allowedIntentPackageNames);
 
             builder.addAction(button.icon(), button.title(), PendingIntent.getActivity(context, msg.hashCode() + i + 1, pushActivityIntent, Build.VERSION.SDK_INT >= 23 ? PendingIntent.FLAG_IMMUTABLE : 0));
         }
@@ -431,7 +426,16 @@ public class CountlyPush {
         }
 
         return Boolean.TRUE;
+    }
 
+    private static Intent createPushActivityIntent(final Context context, final Message msg, final Intent notificationIntent, int index, Set<String> allowedIntentClassNames, Set<String> allowedIntentPackageNames) {
+        Intent pushActivityIntent = new Intent(context.getApplicationContext(), CountlyPushActivity.class)
+            .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+        pushActivityIntent.setPackage(context.getApplicationContext().getPackageName());
+        pushActivityIntent.putExtra(EXTRA_INTENT, actionIntent(context, notificationIntent, msg, index));
+        pushActivityIntent.putStringArrayListExtra(ALLOWED_CLASS_NAMES, new ArrayList<>(allowedIntentClassNames));
+        pushActivityIntent.putStringArrayListExtra(ALLOWED_PACKAGE_NAMES, new ArrayList<>(allowedIntentPackageNames));
+        return pushActivityIntent;
     }
 
     private static Intent actionIntent(Context context, Intent notificationIntent, Message message, int index) {
@@ -624,8 +628,8 @@ public class CountlyPush {
             Countly.sharedInstance().L.i("[CountlyPush, onTokenRefresh] Consent not given, ignoring call");
             return;
         }
-        Countly.sharedInstance().L.i("[CountlyPush, onTokenRefresh] Refreshing FCM push token, with mode: [" + mode + "] for [" + provider + "]");
-        Countly.sharedInstance().onRegistrationId(token, mode, provider);
+        Countly.sharedInstance().L.i("[CountlyPush, onTokenRefresh] Refreshing FCM push token, with mode: [" + countlyConfigPush.mode + "] for [" + provider + "]");
+        Countly.sharedInstance().onRegistrationId(token, countlyConfigPush.mode, provider);
     }
 
     static final String FIREBASE_MESSAGING_CLASS = "com.google.firebase.messaging.FirebaseMessaging";
@@ -641,6 +645,7 @@ public class CountlyPush {
      * @param application application instance
      * @param mode whether to mark push token as test or as production one
      * @throws IllegalStateException
+     * @deprecated use 'CountlyConfigPush' object to init Countly Push: 'init(CountlyConfigPush countlyConfigPush)'.
      */
     public static void init(final Application application, Countly.CountlyMessagingMode mode) throws IllegalStateException {
         init(application, mode, null);
@@ -653,43 +658,55 @@ public class CountlyPush {
      * @param mode whether to mark push token as test or as production one
      * @param preferredProvider prefer specified push provider, {@code null} means use FCM first, then fallback to Huawei
      * @throws IllegalStateException
+     * @deprecated use 'CountlyConfigPush' object to init Countly Push: 'init(CountlyConfigPush countlyConfigPush)'.
      */
     public static void init(final Application application, Countly.CountlyMessagingMode mode, Countly.CountlyMessagingProvider preferredProvider) throws IllegalStateException {
-        Countly.sharedInstance().L.i("[CountlyPush, init] Initialising Countly Push, App:[" + (application != null) + "], mode:[" + mode + "] provider:[" + preferredProvider + "]");
+        CountlyConfigPush countlyConfigPush = new CountlyConfigPush(application, mode)
+            .setProvider(preferredProvider);
+        init(countlyConfigPush);
+    }
 
-        if (application == null) {
+    /**
+     * Initialize Countly messaging functionality
+     *
+     * @param countlyConfigPush push configuration
+     * @throws IllegalStateException
+     */
+    public static void init(CountlyConfigPush countlyConfigPush) throws IllegalStateException {
+        Countly.sharedInstance().L.i("[CountlyPush, init] Initialising Countly Push, App:[" + (countlyConfigPush.application != null) + "], mode:[" + countlyConfigPush.mode + "] provider:[" + countlyConfigPush.provider + "]");
+
+        if (countlyConfigPush.application == null) {
             throw new IllegalStateException("Non 'null' application must be provided!");
         }
 
+        CountlyPush.countlyConfigPush = countlyConfigPush;
         // set preferred push provider
-        CountlyPush.provider = preferredProvider;
-        if (provider == null) {
+        if (countlyConfigPush.provider == null) {
             if (UtilsMessaging.reflectiveClassExists(FIREBASE_MESSAGING_CLASS, Countly.sharedInstance().L)) {
-                provider = Countly.CountlyMessagingProvider.FCM;
+                countlyConfigPush.setProvider(Countly.CountlyMessagingProvider.FCM);
             } else if (UtilsMessaging.reflectiveClassExists(HUAWEI_MESSAGING_CLASS, Countly.sharedInstance().L)) {
-                provider = Countly.CountlyMessagingProvider.HMS;
+                countlyConfigPush.setProvider(Countly.CountlyMessagingProvider.HMS);
             }
-        } else if (provider == Countly.CountlyMessagingProvider.FCM && !UtilsMessaging.reflectiveClassExists(FIREBASE_MESSAGING_CLASS, Countly.sharedInstance().L)) {
-            provider = Countly.CountlyMessagingProvider.HMS;
-        } else if (provider == Countly.CountlyMessagingProvider.HMS && !UtilsMessaging.reflectiveClassExists(HUAWEI_MESSAGING_CLASS, Countly.sharedInstance().L)) {
-            provider = Countly.CountlyMessagingProvider.FCM;
+        } else if (countlyConfigPush.provider == Countly.CountlyMessagingProvider.FCM && !UtilsMessaging.reflectiveClassExists(FIREBASE_MESSAGING_CLASS, Countly.sharedInstance().L)) {
+            countlyConfigPush.setProvider(Countly.CountlyMessagingProvider.HMS);
+        } else if (countlyConfigPush.provider == Countly.CountlyMessagingProvider.HMS && !UtilsMessaging.reflectiveClassExists(HUAWEI_MESSAGING_CLASS, Countly.sharedInstance().L)) {
+            countlyConfigPush.setProvider(Countly.CountlyMessagingProvider.FCM);
         }
 
         // print error in case preferred push provider is not available
-        if (provider == Countly.CountlyMessagingProvider.FCM && !UtilsMessaging.reflectiveClassExists(FIREBASE_MESSAGING_CLASS, Countly.sharedInstance().L)) {
+        if (countlyConfigPush.provider == Countly.CountlyMessagingProvider.FCM && !UtilsMessaging.reflectiveClassExists(FIREBASE_MESSAGING_CLASS, Countly.sharedInstance().L)) {
             Countly.sharedInstance().L.e("Countly push didn't initialise. No FirebaseMessaging class in the class path. Please either add it to your gradle config or don't use CountlyPush.");
             return;
-        } else if (provider == Countly.CountlyMessagingProvider.HMS && !UtilsMessaging.reflectiveClassExists(HUAWEI_MESSAGING_CLASS, Countly.sharedInstance().L)) {
+        } else if (countlyConfigPush.provider == Countly.CountlyMessagingProvider.HMS && !UtilsMessaging.reflectiveClassExists(HUAWEI_MESSAGING_CLASS, Countly.sharedInstance().L)) {
             Countly.sharedInstance().L.e("Countly push didn't initialise. No HmsMessageService class in the class path. Please either add it to your gradle config or don't use CountlyPush.");
             return;
-        } else if (provider == null) {
+        } else if (countlyConfigPush.provider == null) {
             Countly.sharedInstance().L.e("Countly push didn't initialise. Neither FirebaseMessaging, nor HmsMessageService class in the class path. Please either add Firebase / Huawei dependencies or don't use CountlyPush.");
             return;
         }
 
-        CountlyPush.mode = mode;
-        CountlyStore.cacheLastMessagingMode(mode == Countly.CountlyMessagingMode.TEST ? 0 : 1, application);
-        CountlyStore.storeMessagingProvider(provider == Countly.CountlyMessagingProvider.FCM ? 1 : 2, application);
+        CountlyStore.cacheLastMessagingMode(countlyConfigPush.mode == Countly.CountlyMessagingMode.TEST ? 0 : 1, countlyConfigPush.application);
+        CountlyStore.storeMessagingProvider(countlyConfigPush.provider == Countly.CountlyMessagingProvider.FCM ? 1 : 2, countlyConfigPush.application);
 
         if (callbacks == null) {
             callbacks = new Application.ActivityLifecycleCallbacks() {
@@ -726,21 +743,21 @@ public class CountlyPush {
                 }
             };
 
-            application.registerActivityLifecycleCallbacks(callbacks);
+            countlyConfigPush.application.registerActivityLifecycleCallbacks(callbacks);
 
             IntentFilter filter = new IntentFilter();
             filter.addAction(Countly.CONSENT_BROADCAST);
             BroadcastReceiver consentReceiver = new ConsentBroadcastReceiver();
-            application.registerReceiver(consentReceiver, filter, application.getPackageName() + COUNTLY_BROADCAST_PERMISSION_POSTFIX, null);
+            countlyConfigPush.application.registerReceiver(consentReceiver, filter, countlyConfigPush.application.getPackageName() + COUNTLY_BROADCAST_PERMISSION_POSTFIX, null);
         }
 
-        if (provider == Countly.CountlyMessagingProvider.HMS && getPushConsent(application)) {
+        if (countlyConfigPush.provider == Countly.CountlyMessagingProvider.HMS && getPushConsent(countlyConfigPush.application)) {
             String version = getEMUIVersion();
             if (version.startsWith("10")) {
                 new Thread(new Runnable() {
                     @Override
                     public void run() {
-                        String token = getToken(application, Countly.CountlyMessagingProvider.HMS, Countly.sharedInstance().L);
+                        String token = getToken(countlyConfigPush.application, Countly.CountlyMessagingProvider.HMS, Countly.sharedInstance().L);
                         if (token != null && !"".equals(token)) {
                             onTokenRefresh(token, Countly.CountlyMessagingProvider.HMS);
                         }